% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broken_stick.R
\name{broken_stick}
\alias{broken_stick}
\title{Broken-stick (segmented) approaches to estimating SM50}
\usage{
broken_stick(
  df,
  x,
  y,
  verbose = FALSE,
  ci = 95,
  lower = stats::quantile(x, 0.2),
  upper = stats::quantile(x, 0.8),
  scale = c("log", "none", "std"),
  method = c("segmented", "chngpt", "regrans", "stevens", "all")
)
}
\arguments{
\item{df}{data frame, matrix, or tibble containing the data}

\item{x}{Integer or double vector of measurements for the x-axis variable
(e.g., carapace width).}

\item{y}{Integer or double vector of measurements for the y-axis variable
(e.g., claw height).}

\item{verbose}{Boolean; Should the standard error, confidence intervals,}

\item{ci}{Integer; type of confidence intervals to return for SM50, defaults
to 95\%.}

\item{lower}{Integer or double; the lower bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the w0th percentile of
the x-variable.}

\item{upper}{Integer or double; the upper bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the 80th percentile of
the x-variable.}

\item{scale}{Transformation to be applied to the data before performing the
regression: "none", "log" (both variables are log-transformed), or "std"
(both variables are standardized = scaled and centered). If no string is
provided, no transformation is performed (i.e., the default is "none").}

\item{method}{Method to use for the regression. A single string or string
vector containing one or more of c("segmented", "chngpt", "regrans",
"stevens"), or "all" to return the results of all methods for comparison.}
}
\value{
If verbose is FALSE (the default), an estimate of SM50 from the
specified method(s). Otherwise, output is a list that also includes the
standard error and confidence intervals for the SM50 estimate as well as
the model object(s) and/or regression parameters for the lines in the
broken-stick models.
}
\description{
A wrapper function allowing for multiple methods of broken-stick regression to be applied using a standard format for inputs. See the `vignette("Broken-stick methods") vignette for more information.
}
\examples{
broken_stick(df=iris, x="x", y="y", method=c("segmented", "chngpt"))
}
