% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_mods}
\alias{cluster_mods}
\title{Classification approaches to estimating SM50}
\usage{
cluster_mods(
  dat,
  xvar,
  yvar,
  log = FALSE,
  method = c("mclust", "Somerton", "kmeans", "hclust", "infl_pt", "all")
)
}
\arguments{
\item{dat}{data frame or matrix containing the data}

\item{xvar}{Name of column (integer or double) of measurements for the x-axis
variable (e.g., carapace width).}

\item{yvar}{Name of column (integer or double) of measurements for the y-axis
variable (e.g., claw height).}

\item{log}{Boolean; should both variables be log-transformed before performing the
regression? Defaults to FALSE.}

\item{method}{Classification method to use. A single string or vector
containing one or more of c("mclust", "Somerton", "kmeans", "hclust", "infl_pt"), or "all" to return the results of all methods for comparison.}
}
\value{
An estimate of SM50 from the specified method(s).
}
\description{
Classification approaches to estimating SM50
}
\examples{
set.seed(12)
fc <- fake_crustaceans(n = 100, L50 = 100, allo_params = c(1, 0.2, 1.1, 0.2))
cluster_mods(fc, xvar = "x", yvar = "y", method = c("kmeans"))
}
