% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_int.R
\name{density_int}
\alias{density_int}
\title{Distribution interval derived (DID) cutline method}
\usage{
density_int(
  dat,
  xvar,
  yvar,
  lower = NULL,
  upper = NULL,
  int_num = 25,
  log = FALSE,
  plot = FALSE
)
}
\arguments{
\item{dat}{data frame or matrix containing the data}

\item{xvar}{Name of column (or integer or double vector) containing
measurements for the x-axis variable (e.g., carapace width).}

\item{yvar}{Name of column (or integer or double vector) containing
measurements for the y-axis variable (e.g., claw height).}

\item{lower}{Integer or double; the lower bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the 20th percentile of
the x-variable.}

\item{upper}{Integer or double; the upper bound for possible SM50 values.
Must be on the same scale as the data. Defaults to the 80th percentile of
the x-variable.}

\item{int_num}{Integer; how many intervals between the lower and upper bound
should be used? Defaults to 25. With fewer intervals, each interval will
contain more points, increasing the accuracy of the estimated density
minimum for a given interval. However, the linear regression of the minima
distributions (the divisions between immature and mature individuals within
an interval) against the midpoints of those intervals may be more reliable
with more intervals.}

\item{log}{Boolean; should both variables be log-transformed before
performing the regression? Defaults to FALSE.}

\item{plot}{Boolean; should a plot of the data with the calculated minima and
discriminating line be displayed?}
}
\value{
Something
}
\description{
Distribution interval derived (DID) cutline method
}
\examples{
set.seed(12)
fc <- fake_crustaceans(n = 1000, L50 = 100, allo_params = c(1, 0.2, 1.1, 0.2))
density_int(dat = fc, xvar = "x", yvar = "y", upper = 120)
density_int(dat = fc, xvar = "x", yvar = "y", upper = log(120), log = TRUE)
}
