% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_crustaceans.R
\name{fake_crustaceans}
\alias{fake_crustaceans}
\title{Generate artificial data for simulation testing}
\usage{
fake_crustaceans(
  L50 = 100,
  slope = 5,
  n = 1000,
  x_mean = 105,
  x_sd = 20,
  allo_params = c(1.2, 0.1, 1.2, 0.1),
  error_scale = 20
)
}
\arguments{
\item{L50}{Integer or double; the desired true length at 50\% maturity on the
scale of the x-axis/reference variable. Defaults to 100 mm.}

\item{slope}{Integer or double; the desired slope parameter for the logistic
equation describing the probability of maturity at a given value of the
x-axis/reference variable. Default is 5.}

\item{n}{Sample size of the simulated data set. Default is 1000 individuals.}

\item{x_mean}{Mean of the reference variable (e.g., carapace width). Default
is 105 mm.}

\item{x_sd}{Standard deviation of the reference variable (e.g., carapace
width). Default is 20 mm.}

\item{allo_params}{A numeric vector of length 4 containing the parameters
controlling how the allometric relationship between the x and y variables
changes at maturity. Should contain the immature slope parameter, immature
intercept parameter, mature slope parameter, and mature intercept
parameter, in that order.}

\item{error_scale}{Scaling for the error added to the simulated data}
}
\value{
A data frame with n rows. Columns are: (1) the x variable on the
original scale, (2) the probability of maturity for the individual, (3) the
assigned maturity status, 1 or 0, (4) the y variable on the original scale,
(5) the log-transformed x variable, and (6) the log-transformed y variable
}
\description{
Generates an artificial sample of morphometric data with
specified characteristics. Recommended to use set.seed() before running
to ensure reproducibility.
}
\examples{
set.seed(123)
fake_crustaceans(n=25)
}
