% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infl_pt.R
\name{infl_pt_fun}
\alias{infl_pt_fun}
\title{Maturity classification based on the minimum density of CH/CW ratios}
\usage{
infl_pt_fun(x, y, plot = FALSE)
}
\arguments{
\item{x}{Integer or double vector of measurements for the x-axis
variable (e.g., carapace width).}

\item{y}{Integer or double vector of measurements for the y-axis
variable (e.g., claw height).}

\item{plot}{Boolean; should a plot of the density curve with the identified
minimum be created?}
}
\value{
An integer corresponding to the ratio (y/x) at which the local
minimum was found. This ratio serves as the discriminant line when
separating observations by maturity status.
}
\description{
Maturity classification based on the minimum density of CH/CW ratios
}
\examples{
library(ggplot2)
set.seed(12)
x <- rnorm(100, mean=2, sd=3)
y <- rnorm(100, mean=10, sd=3)
z <- c(x, y)
hist(z)
infl_pt_fun(1, z, TRUE)

}
