% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphmat-package.R
\docType{package}
\name{morphmat-package}
\alias{morphmat}
\alias{morphmat-package}
\title{morphmat: Modeling Size at Maturity Using Morphometric Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A compilation of methods used to estimate size at (sexual) maturity based on morphometric data, most commonly applied to crabs, lobsters, and other crustaceans. Approaches include modeling approaches based on piecewise/segmentedlinear regression as well as numerous clustering-based methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmk118/morphmat}
  \item Report bugs at \url{https://github.com/rmk118/morphmat/issues}
}

}
\author{
\strong{Maintainer}: Ruby Krasnow \email{ruby.krasnow@maine.edu} (\href{https://orcid.org/0009-0007-0531-8420}{ORCID})

}
\keyword{internal}
