% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise_mods}
\alias{piecewise_mods}
\title{Piecewise regression approaches to estimating SM50}
\usage{
piecewise_mods(
  dat,
  xvar,
  yvar,
  lower = NULL,
  upper = NULL,
  log = FALSE,
  method = c("segmented", "chngpt", "regrans", "stevens", "TL", "all")
)
}
\arguments{
\item{dat}{data frame or matrix containing the data}

\item{xvar}{Name of column (integer or double) of measurements for the x-axis
variable (e.g., carapace width).}

\item{yvar}{Name of column (integer or double) of measurements for the y-axis
variable (e.g., claw height).}

\item{lower}{Integer or double; the lower bound for possible SM50 values.
Must be on the same trans of the data. Defaults to the 20th percentile of
the x-variable.}

\item{upper}{Integer or double; the upper bound for possible SM50 values.
Must be on the same trans of the data. Defaults to the 80th percentile of
the x-variable.}

\item{log}{Boolean; should both variables be log-transformed before performing the
regression? Defaults to FALSE.}

\item{method}{Method to use for the regression. A single string or vector
containing one or more of c("segmented", "chngpt", "regrans", "stevens",
"TL"), or "all" to return the results of all methods for comparison.}
}
\value{
An estimate of SM50 from the specified method(s).
}
\description{
A wrapper function allowing for multiple methods of piecewise regression to
be applied using a standard format for inputs. See \code{vignette("broken-stick")}
and \code{vignette("two-line")} for more information.
}
\details{
The \code{two_line_logistic()} function is closely related but not included in this wrapper function because you will generally want more control over the initial values of the parameters and the nonlinear least-squares algorithm may not always converge.

This function is primarily intended for easy comparison between the SM50 estimates produced by a variety of different piecewise regression models. For follow-up analyses, we recommend calling the specific function(s) of interest (\code{regrans()}, \code{broken_stick_stevens()}, \code{two_line()}, \code{segmented::segmented()}, or \code{chngpt::chngpt()}) and exploring how changing the upper and lower bounds for possible SM50 values and the number of breakpoints to be tested may influence the resulting estimates.
}
\examples{
set.seed(12)
fc <- fake_crustaceans(n = 100, L50 = 100, allo_params = c(1, 0.2, 1.1, 0.2))
piecewise_mods(fc, xvar = "x", yvar = "y", method = c("segmented", "chngpt"))
}
