% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrans.R
\name{regrans}
\alias{regrans}
\title{REGRANS broken-stick regression method}
\usage{
regrans(
  dat,
  xvar,
  yvar,
  lower = NULL,
  upper = NULL,
  log = FALSE,
  verbose = FALSE,
  n_tries = 100
)
}
\arguments{
\item{dat}{data frame or matrix containing the data}

\item{xvar}{Name of column (integer or double) of measurements for the x-axis
variable (e.g., carapace width).}

\item{yvar}{Name of column (integer or double) of measurements for the y-axis
variable (e.g., claw height).}

\item{lower}{Integer or double; the lower bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the 20th percentile of
the x-variable.}

\item{upper}{Integer or double; the upper bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the 80th percentile of
the x-variable.}

\item{log}{Boolean; should both variables be log-transformed before performing the
regression? Defaults to FALSE.}

\item{verbose}{Return all breakpoints tested and their sum of squares, or
only the estimated SM50?}

\item{n_tries}{Number of breakpoints to test within the unknown range.}
}
\value{
If verbose = TRUE, a data frame with the breakpoints tested and
their sum of squares. Otherwise, a single value for the breakpoint with the
lowest sum of squares.
}
\description{
REGRANS broken-stick regression method
}
\examples{
set.seed(12)
fc <- fake_crustaceans(n = 100, L50 = 100, allo_params = c(1, 0.2, 1.1, 0.2))
regrans(fc, "x", "y", verbose = FALSE)
head(regrans(fc, "x", "y", verbose = TRUE), n = 30)
}
