% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrans.R
\name{regrans_fun}
\alias{regrans_fun}
\title{REGRANS broken-stick regression method}
\usage{
regrans_fun(
  x,
  y,
  min_changept = min(x, na.rm = TRUE),
  max_changept = max(x, na.rm = TRUE),
  n_tries = 100
)
}
\arguments{
\item{x}{Integer or double vector of measurements for the x-axis variable
(e.g., carapace width).}

\item{y}{Integer or double vector of measurements for the y-axis variable
(e.g., claw height).}

\item{min_changept}{Integer or double; the lower bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the minimum value of the x-variable.}

\item{max_changept}{Integer or double; the upper bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the maximum value of
the x-variable.}

\item{n_tries}{Number of breakpoints to test within the unknown range.}
}
\value{
Data frame with breakpoints tested and their sum of squares
}
\description{
REGRANS broken-stick regression method
}
\examples{
fc <- fake_crabs(n=100, allo_params=c(1, 0.2, 1.1, 0.2))
regrans_fun(fc$x, fc$y)
}
