% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somerton.R
\name{somerton_fun}
\alias{somerton_fun}
\title{Somerton method}
\usage{
somerton_fun(
  dat,
  xvar,
  yvar,
  trans = "none",
  lower = NULL,
  upper = NULL,
  max_iter = 50
)
}
\arguments{
\item{dat}{data frame or matrix containing the data}

\item{xvar}{Name of column (integer or double) of measurements for the x -
axis variable (e.g., carapace width).}

\item{yvar}{Name of column (integer or double) of measurements for the y-axis
variable (e.g., claw height).}

\item{trans}{Transformation to be applied to the data before performing the
regression: "none", "log" (both variables are log-transformed), or "std"
(both variables are standardized = scaled and centered). If no string is
provided, no transformation is performed (i.e., the default is "none").}

\item{lower}{Integer or double; the lower bound for possible SM50 values.
Must be on the same trans of the data. Defaults to the 20th percentile of
the x-variable.}

\item{upper}{Integer or double; the upper bound for possible SM50 values.
Must be on the same trans of the data. Defaults to the 80th percentile of
the x-variable.}

\item{max_iter}{Maximum number of iterations}
}
\value{
Output is a list that also includes the input data frame with a
column specifying which maturity group each point was assigned to, vectors
of the R-squared and residual sum of squares for each iteration, and the
linear model objects corresponding to each maturity group.
}
\description{
Somerton method
}
\examples{
set.seed(12)
fc <- fake_crustaceans(n = 100, L50 = 100, allo_params = c(1, 0.2, 1.1, 0.2))
out_df <- somerton_fun(fc, xvar = "x", yvar = "y")[[1]]
mod <- glm(data = out_df, pred_mat_num ~ x, family = binomial(link = "logit"))
unname(-coef(mod)[1] / coef(mod)[2])
}
