% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_line_logistic.R
\name{two_line_logistic}
\alias{two_line_logistic}
\title{Two-line logistic model}
\usage{
two_line_logistic(
  dat,
  xvar,
  yvar,
  imm_int = 1,
  imm_slope = 0.2,
  mat_int = 1,
  mat_slope = 0.3,
  SM50_start = NULL,
  alpha_start = 9,
  verbose = FALSE
)
}
\arguments{
\item{dat}{data frame or matrix containing the data}

\item{xvar}{Name of column (integer or double) of measurements for the x-axis
variable (e.g., carapace width).}

\item{yvar}{Name of column (integer or double) of measurements for the y-axis
variable (e.g., claw height).}

\item{imm_int}{Starting value for the immature intercept parameter when
fitting the NLS model}

\item{imm_slope}{tarting value for the immature slope parameter when fitting
the NLS model}

\item{mat_int}{Starting value for the mature intercept parameter when fitting
the NLS model}

\item{mat_slope}{Starting value for the mature slope parameter when fitting
the NLS model}

\item{SM50_start}{Starting value for SM50 parameter when fitting the NLS
model. If not provided, taken to be the median of the x-variable}

\item{alpha_start}{Starting value for the logistic slope parameter when
fitting the NLS model}

\item{verbose}{Should additional output be returned besides the SM50
estimate?}
}
\value{
If verbose is FALSE (the default), an estimate of SM50. Otherwise,
output is the NLS model object.
}
\description{
Two-line logistic model
}
\examples{
set.seed(12)
fc <- fake_crustaceans(n=100, L50=100, allo_params=c(1, 0.2, 1.1, 0.2))
two_line_logistic(fc, xvar="x", yvar="y", verbose = FALSE)
}
