% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_line_stevens.R
\name{two_line_stevens}
\alias{two_line_stevens}
\title{Two-line methods from Bradley Stevens}
\usage{
two_line_stevens(
  dat,
  xvar,
  yvar,
  lower = NULL,
  upper = NULL,
  verbose = FALSE,
  bps = "even",
  num_bps = 100
)
}
\arguments{
\item{dat}{data frame or matrix containing the data}

\item{xvar}{Name of column (integer or double) of measurements for the x-axis
variable (e.g., carapace width).}

\item{yvar}{Name of column (integer or double) of measurements for the y-axis
variable (e.g., claw height).}

\item{lower}{Integer or double; the lower bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the 20th percentile of
the x-variable.}

\item{upper}{Integer or double; the upper bound for possible SM50 values.
Must be on the same scale of the data. Defaults to the 80th percentile of
the x-variable.}

\item{verbose}{Should additional output be returned besides the SM50
estimate?}

\item{bps}{Should the values tested as possible breakpoints be restricted to
only observed values of the x-variable ("obs"), or should it be a specified
number of evenly-spaced values between the lower and upper limits of the
unknown region ("even", the default)}

\item{num_bps}{When \code{bps = "even"}, how many values should be tested as
possible endpoints? Defaults to 100, but should be increased.}
}
\value{
If verbose is FALSE (the default), two possible estimates of SM50:
the breakpoint x-value marking the transition between immature and mature
points/lines, and the intersection point where the two lines cross. The
intersection value will typically be extremely unrealistic unless
the slopes of the lines are drastically different. If verbose is TRUE,
output is a list that also includes the original data with a column
representing which line (immature or mature) the point was assigned to, the
immature amd mature slope and intercept parameters, and the intersection
point of the two lines.
}
\description{
Two-line methods from Bradley Stevens
}
\examples{
#' set.seed(12)
fc <- fake_crustaceans(n = 100, L50 = 100, allo_params = c(1, 0.2, 1.1, 0.2))
two_line_stevens(fc, xvar = "x", yvar = "y", verbose = FALSE)
}
