% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broken_stick.R
\name{broken_stick}
\alias{broken_stick}
\title{Broken-stick (segmented) approaches to estimating SM50}
\usage{
broken_stick(
  dat,
  xvar,
  yvar,
  ci = 95,
  lower = NULL,
  upper = NULL,
  trans = "none",
  method = c("segmented", "chngpt", "regrans", "stevens", "all")
)
}
\arguments{
\item{dat}{data frame or matrix containing the data}

\item{xvar}{Name of column (integer or double) of measurements for the x-axis
variable (e.g., carapace width).}

\item{yvar}{Name of column (integer or double) of measurements for the y-axis
variable (e.g., claw height).}

\item{ci}{Integer; type of confidence intervals to return for SM50, defaults
to 95\%.}

\item{lower}{Integer or double; the lower bound for possible SM50 values.
Must be on the same trans of the data. Defaults to the 20th percentile of
the x-variable.}

\item{upper}{Integer or double; the upper bound for possible SM50 values.
Must be on the same trans of the data. Defaults to the 80th percentile of
the x-variable.}

\item{trans}{Transformation to be applied to the data before performing the
regression: "none", "log" (both variables are log-transformed), or "std"
(both variables are standardized = scaled and centered). If no string is
provided, no transformation is performed (i.e., the default is "none").}

\item{method}{Method to use for the regression. A single string or string
vector containing one or more of c("segmented", "chngpt", "regrans",
"stevens"), or "all" to return the results of all methods for comparison.}
}
\value{
An estimate of SM50 from the specified method(s).
}
\description{
A wrapper function allowing for multiple methods of broken-stick regression
to be applied using a standard format for inputs. See
\code{vignette("broken-stick")} for more information.
}
\examples{
set.seed(12)
fc <- fake_crustaceans(n = 100, L50 = 100, allo_params = c(1, 0.2, 1.1, 0.2))
broken_stick(fc, xvar = "x", yvar = "y", method = c("segmented", "chngpt"))
}
